Mostly bug fixes

+++ analysis/monosynapticPairs/CCG.m
@@ -139,7 +139,7 @@ end
 nEvents = length(times);
 % 
 
-counts = double(CCGHeart(times,uint32(groups),binSize_Fs,uint32(halfBins)));
+counts = double(CCGHeart(double(times),uint32(groups),binSize_Fs,uint32(halfBins)));
 % -----------------------------------
 % 
 % Reshape the results
